requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0.1
pyperclip>=1.8.2
pyinstaller